Fix some fundamental file field issue
File fields can not be assigned to.
That fundamentally does not work as the browser will block you when doing so.
That is because the browser only has permission to give
images out that you selected, it can not access images
by getting values assigned to a field or the like.