Allow lodash and leaflet to be commonjs dependencies
 CommonJS or AMD dependencies can cause optimization issues in the Angular build process,
which may result in slower performance or larger bundle sizes.
This is because Angular uses a modern module system called ES2015 modules,
which are more efficient than CommonJS or AMD modules.

This is unavoidable for leaflet and lodash though.
I don't plan on dropping them, so I shall allow them.