Move all article models back to root
These models are needed in both the services and in detail components.
The issue is that the services in turn are needed in both login and detail module.
So the services need to be on the root level and thus the models too.

Further added some models I hadn't gotten around to adding yet.