Fix forms not reacting properly on enter
https://stackoverflow.com/questions/4763638/enter-triggers-button-click

Turns out that the first button in a form is implicitly "submit".
Pressing enter while an element in a form element is focussed
will trigger the click element on that button.
That causes undesired behaviour,
namely enter presses triggering cancel events.

Now the button explicitly sets the type to not be "submit"
unless otherwise specified.